add: move `read_quantifiers` to `lingo` module
