+++ src/lingo/mod.rs
@@ -1,3 +1,4 @@
 pub mod past_participle;
 pub mod consonants;
 mod verb_conjugation;
+pub mod quantifiers;
