+++ src/wordnet/interface.rs
@@ -78,13 +78,3 @@ pub fn get_meanings(word: &str) -> Option<Vec<Wordnode>> {
     let meanings = WORD_MEANINGS.lock().unwrap();
     meanings.get(word).cloned()
 }
-
-pub fn read_quantifiers() -> io::Result<Vec<String>> {
-    let path = Path::new("data/every_quantifiers.txt");
-    let file = File::open(&path)?;
-    let reader = io::BufReader::new(file);
-
-    let lines: Vec<String> = reader.lines().collect::<Result<_, _>>()?;
-
-    Ok(lines)
-}
